// src/enums/UserRoles.enum.ts
// ✅ User Roles Enum
export enum UserRoles {
    Reporter = 'Reporter',
    DesignerAdPublisher = 'Designer|AD Publisher',
    Updator = 'Updator',
    Moderator = 'Moderator',
    Editor = 'Editor',
    TechSeoAnalyst = 'Tech/SEO Analyst',
    Publisher = 'Publisher',
    SiteAdmin = 'Site Admin',
    SuperAdmin = 'Super Admin'
  }